# Optimization for SpriteBatch when running non VertexArray VertexDataModes. (GL30 default)