--- a/gdx/src/com/badlogic/gdx/graphics/g2d/SpriteBatch.java
+++ b/gdx/src/com/badlogic/gdx/graphics/g2d/SpriteBatch.java
@@ -25,7 +25,6 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.VertexAttribute;
 import com.badlogic.gdx.graphics.VertexAttributes.Usage;
 import com.badlogic.gdx.graphics.glutils.ShaderProgram;
-import com.badlogic.gdx.graphics.glutils.VertexArray;
 import com.badlogic.gdx.math.Affine2;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Matrix4;
@@ -130,7 +129,7 @@ public class SpriteBatch implements Batch {
 		} else
 			shader = defaultShader;
 
-		//Pre bind the mesh to force the upload of indices data.
+		// Pre bind the mesh to force the upload of indices data.
 		if (vertexDataType != VertexDataType.VertexArray) {
 			mesh.bind(shader);
 			mesh.unbind(shader);