In flush of SpriteBatch, the index buffer gets a limit for the current size of the batch. It gets the indices buffer reference whilst marking the index buffer as dirty, so on bind of the mesh, the indices are uploaded again. Since SpriteBatch indices are static and never change, this change pre-uploads all the indices to the IndexBuffer at creation time, and skips any further updates at render.

We don't have a nice way to do this without making a bunch of new methods specifically for this case, so I'm just binding and unbinding the mesh to trigger the upload once before a render occurs.

Testers welcome! Seems to be no regressions for me in the test suite, but would be good if otherr can take a look on their systems and devices


Spritebatch defaults timings from SpriteBatchPerformanceTest
```
before change
gl20 0.225ms
gl30 0.280ms


after change
gl20 0.225ms
gl30 0.201ms 

```


Related issue
https://github.com/libgdx/libgdx/issues/7345