@NathanSweet 2024-04-18
I ran Spine with GL30 with and without these changes, all seems fine. I didn't measure performance differences.

I'm good with merging but I didn't press the button in case obigu's proposed changes will be added.